requirements.txt
# requirements.txt
mysql-connector-python
pyjwt
python-dotenv
azure-functions
sqlalchemy
